rearranged the code somewhat
